+++ src/main/java/page/clab/api/domain/member/application/MemberService.java
@@ -28,9 +28,9 @@ import page.clab.api.domain.position.domain.PositionType;
 import page.clab.api.global.auth.util.AuthUtil;
 import page.clab.api.global.common.dto.PagedResponseDto;
 import page.clab.api.global.common.email.application.EmailService;
-import page.clab.api.global.common.verificationCode.application.VerificationService;
-import page.clab.api.global.common.verificationCode.domain.Verification;
-import page.clab.api.global.common.verificationCode.dto.request.VerificationRequestDto;
+import page.clab.api.global.common.verification.application.VerificationService;
+import page.clab.api.global.common.verification.domain.Verification;
+import page.clab.api.global.common.verification.dto.request.VerificationRequestDto;
 import page.clab.api.global.exception.InvalidInformationException;
 import page.clab.api.global.exception.NotFoundException;
 import page.clab.api.global.exception.PermissionDeniedException;
